Move validateNoLeadingZeroes back to util
